+++ src/main/java/xyz/alycat/hwr/effect/ModStatusEffects.java
@@ -11,7 +11,7 @@ public class ModStatusEffects {
 
     public static StatusEffect registerWaterResistanceEffect(String name) {
         return Registry.register(Registry.STATUS_EFFECT, new Identifier(Hwr.MOD_ID, name),
-                new WaterResistanceEffect(StatusEffectCategory.BENEFICIAL, 1582940));
+                new WaterResistanceEffect(StatusEffectCategory.BENEFICIAL, Hwr.CONFIG.potion_colour().getColour()));
     }
     public static void registerEffects() {
         WATER_RESISTANCE = registerWaterResistanceEffect("water_resistance");