+++ src/main/java/xyz/alycat/hwr/potion/ModPotions.java
@@ -1,6 +1,7 @@
 package xyz.alycat.hwr.potion;
 
 import net.minecraft.entity.effect.StatusEffectInstance;
+import net.minecraft.item.Item;
 import net.minecraft.item.Items;
 import net.minecraft.potion.Potion;
 import net.minecraft.potion.Potions;
@@ -9,10 +10,13 @@ import net.minecraft.util.Identifier;
 import xyz.alycat.hwr.Hwr;
 import xyz.alycat.hwr.effect.ModStatusEffectInstance;
 import xyz.alycat.hwr.mixin.BrewingRecipeRegistryMixin;
+import xyz.alycat.hwr.util.ConfigUtils;
 
 public class ModPotions {
     public static Potion WATER_RESISTANCE;
     public static Potion LONG_WATER_RESISTANCE;
+
+    public static Item POTION_RECIPE = ConfigUtils.getItem(Hwr.CONFIG.potion_recipe(), "minecraft:sponge");
 
     public static Potion registerPotion(String name, StatusEffectInstance statusEffectInstance) {
         return Registry.register(Registry.POTION, new Identifier(Hwr.MOD_ID, name), new Potion(statusEffectInstance));
@@ -25,7 +29,7 @@ public class ModPotions {
     }
 
     private static void registerPotionRecipes() {
-        BrewingRecipeRegistryMixin.invokeRegisterPotionRecipe(Potions.AWKWARD, Items.SPONGE, ModPotions.WATER_RESISTANCE);
+        BrewingRecipeRegistryMixin.invokeRegisterPotionRecipe(Potions.AWKWARD, POTION_RECIPE, ModPotions.WATER_RESISTANCE);
         BrewingRecipeRegistryMixin.invokeRegisterPotionRecipe(ModPotions.WATER_RESISTANCE, Items.REDSTONE, ModPotions.LONG_WATER_RESISTANCE);
     }
 }