+++ docs/CHANGELOG.md
@@ -1,3 +1,29 @@
+### 1.3.0 - 2024.02.18
+
+*   Changed default potion colour
+*   Added configuration options for potion colour
+*   Potion recipe option no longer unnecessarily forces sync upon joining server
+*   Better dependency version requirements 
+    *   (allow same version of mod to work on more Minecraft versions)
+*   Localisation changes
+
+
+### 1.2.0 - 2024.02.14
+
+*   Added configuration options for potion recipe
+*   Use newer build tools
+
+
+### 1.1.1 - 2023.11.30
+
+
+*   Enchanted golden apples now always apply the water resistance effect upon consumption instead of only giving it once
+
+\
+*   Technical changes:
+    * Rebuilt with updated dependencies
+
+
 ### 1.1.0 - 2023.07.22
 
 *	Enchanted golden apples now give water resistance effect